Removes custom group delete action as obsolete
The custom action made sure that group icons get removed
along with the group despite the fact that the icons are not
stored under the group but the group owner entity.

This feature is nowadays handled correctly by the Elgg core
so there's no use for the custom action anymore.